Initial Commit
The ability to generate characters is all there already, with a decent amount of options and a handful of (physical based) classes.

I want to get the game playable before we bother adding magical stuff in.

A few basic weapons, armor, and potions are there too - we just need some random misc items.