Generate Dungeons Complete
Can now generate dungeons, and the starting room *should* always be connected to the boss room. Just in case it's not, we should probably give the player the ability to exit the dungeon at any time. Maybe using an item?

Also cleaned up the code a little with a few comments.