A Large Chunk of Combat
Made it so characters can be saved and loaded, so one character can persist though multiple rounds of the combat test now.

Implemented a potion of the Inventory, and spent a lot of time to make the code more consistent throughout.

Added formulas for EXP required to level up, as well as implemented proficiency calculationg with formula.

Added an option for seeing weight/height in in metric or imperial scale.

Added typewriter effect for printing text out - also loads from the settings for the delay.